converted Xtend test to Java
